+++ __tests__/__mocks__/data/mockHomePageMainContent.tsx
@@ -1,9 +1,9 @@
 import { imageURL } from "../../../app/src/utils/utils";
-import { mockSwimLanes } from "./mockSwimLanes";
+import { mockCollectionLanes } from "./mockCollectionLanes";
 
 export const mockHomePageMainContent = {
-  swimLaneData: {
-    lanesWithNumItems: mockSwimLanes.lanesWithNumItems,
+  laneData: {
+    lanesWithNumItems: mockCollectionLanes.lanesWithNumItems,
     randomNumber: 1,
   },
   featuredItemData: {