+++ app/src/components/pageLayout/pageLayout.tsx
@@ -85,7 +85,8 @@ const PageLayout = ({
         <Header />
         {activePage === "home" ||
         activePage === "about" ||
-        activePage === "notFound" ? (
+        activePage === "notFound" ||
+        activePage === "serverError" ? (
           children
         ) : (
           <>